mp4 sending

+++ bot.py
@@ -1,5 +1,4 @@
 from aiogram import Bot, Dispatcher, executor, types
-from asyncio import sleep
 from dotenv import load_dotenv
 from io import BytesIO
 
@@ -8,6 +7,7 @@ import os
 import re
 import requests
 import json
+import random
 
 load_dotenv()
 bot = Bot(token=os.getenv('TOKEN'))
@@ -78,8 +78,39 @@ async def current_message(message: types.message):
         await bot.send_message(message.from_user.id, "No channel connected! Forward me any message from it!")
 
 
-@dp.message_handler(content_types=["text"])
-async def setup2_message(message: types.message):
+# @dp.message_handler(content_types=["text"])
+# async def setup2_message(message: types.message):
+#     global channelid
+#     global channelname
+#
+#     if message.forward_from_chat:
+#         if not message.forward_from_chat.id:
+#             await bot.send_message(message.from_user.id, "I need a message forwarded from your channel!")
+#         else:
+#             with open('channels.json') as f:
+#                 data = json.load(f)
+#                 isnew = True
+#                 for u in data['users']:
+#                     if message.from_user.id == u['id']:
+#                         u['channelid'] = message.forward_from_chat.id
+#                         u['channelname'] = message.forward_from_chat.username
+#                         isnew = False
+#             with open('channels.json', 'w') as f:
+#                 if isnew:
+#                     newuser = ({'id': message.from_user.id, 'channelid': message.forward_from_chat.id,
+#                                 'channelname': message.forward_from_chat.username})
+#                     data['users'].append(newuser)
+#                 json.dump(data, f)
+#                 f.close()
+#         channelid = message.forward_from_chat.id
+#         channelname = message.forward_from_chat.username
+#         await bot.send_message(message.from_user.id,
+#                                "Ready to post to channel: " + f'@{channelname}' + " Don't forget that i must be the admin of this channel!")
+
+
+@dp.message_handler(content_types=['document', 'text'])
+async def convert_webm(message: types.file):
+    global ffm
     global channelid
     global channelname
 
@@ -107,41 +138,36 @@ async def setup2_message(message: types.message):
         await bot.send_message(message.from_user.id,
                                "Ready to post to channel: " + f'@{channelname}' + " Don't forget that i must be the admin of this channel!")
 
-
-@dp.message_handler(content_types=['document', 'text'])
-async def convert_webm(message: types.file):
-    global ffm
-    global channelid
-
-    if channelid == []:
+    if channelid == 0:
         await bot.send_message(message.from_user.id, "First you need to setup your server!")
-    elif message.document.mime_subtype == 'webm':
-        try:
-            webmid = message.document.file_id
-            await bot.send_message(message.from_user.id, "Converting to mp4 ...")
-            webmvid = await bot.download_file_by_id(webmid)
+    elif message.document:
+        if message.document.mime_subtype == 'webm':
+            try:
+                webmid = message.document.file_id
+                await bot.send_message(message.from_user.id, "Converting to mp4 ...")
+                webmvid = await bot.download_file_by_id(webmid)
 
-            input_file_name = str(webmid + 'input.webm')
-            output_file_name = str(webmid + 'output.mp4')
+                input_file_name = str(webmid + 'input.webm')
+                output_file_name = str(webmid + 'output.mp4')
 
-            b = BytesIO()
-            b.write(webmvid.getvalue())
-            with open(input_file_name, 'wb') as f:
-                f.write(b.getvalue())
-            ffm.convert_webm_mp4(input_file_name, output_file_name)
-            file = types.InputFile(output_file_name)
-            await bot.send_document(int(channelid), file)
-            path = os.path.join(os.path.abspath(os.path.dirname(__file__)), input_file_name)
-            os.remove(path)
-            path = os.path.join(os.path.abspath(os.path.dirname(__file__)), output_file_name)
-            os.remove(path)
-        except:
-            path = os.path.join(os.path.abspath(os.path.dirname(__file__)), input_file_name)
-            os.remove(path)
-            await bot.send_message(message.from_user.id, "Conversion error!")
+                b = BytesIO()
+                b.write(webmvid.getvalue())
+                with open(input_file_name, 'wb') as f:
+                    f.write(b.getvalue())
+                ffm.convert_webm_mp4(input_file_name, output_file_name)
+                file = types.InputFile(output_file_name)
+                await bot.send_document(int(channelid), file)
+                path = os.path.join(os.path.abspath(os.path.dirname(__file__)), input_file_name)
+                os.remove(path)
+                path = os.path.join(os.path.abspath(os.path.dirname(__file__)), output_file_name)
+                os.remove(path)
+            except:
+                path = os.path.join(os.path.abspath(os.path.dirname(__file__)), input_file_name)
+                os.remove(path)
+                await bot.send_message(message.from_user.id, "Conversion error!")
 
     elif re.search(r'webm$', message.text.lower()):
-        webmid = 'dsfghsdgdfsg'  # ЭТО КЕК !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+        webmid = str(random.randrange(10000000))
         await bot.send_message(message.from_user.id, "Converting to mp4 ...")
         request = requests.get(message.text)
         input_file_name = str(webmid + 'input.webm')
@@ -158,13 +184,16 @@ async def convert_webm(message: types.file):
         os.remove(path)
 
     elif re.search(r'mp4$', message.text.lower()):
-        webmid = 'dsfghsdgdfsg'  # ЭТО КЕК !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-        await bot.send_message(message.from_user.id, "Converting to mp4 ...")
+        webmid = str(random.randrange(10000000))
         request = requests.get(message.text)
         file_name = str(webmid + 'vid.mp4')
+        await bot.send_message(message.from_user.id, "Downloading mp4")
         with open(file_name, 'wb') as file:
             file.write(request.content)
-        await bot.send_video(int(channelid), file, supports_streaming=True)
+            file.close()
+        await bot.send_message(message.from_user.id, "Sending mp4")
+        file = types.InputFile(file_name)
+        await bot.send_video(int(channelid), file)
 
         path = os.path.join(os.path.abspath(os.path.dirname(__file__)), file_name)
         os.remove(path)
